requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
drf-yasg>=1.21.7

cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
psycopg2-binary>=2.9.9

gunicorn>=21.2.0
whitenoise>=6.6.0
